app/requirements/base.txt
Django>=1.11,<2.0
psycopg2-binary
celery>=4.2,<4.3
djangorestframework==3.7.3
pytz==2017.3
djangorestframework-simplejwt
django-cors-headers
